完成-创建表, 插入或更新数据(V2版本)

+++ SJDBMap/Model/SJDBMapQueryCache.m
@@ -33,7 +33,7 @@
 }
 
 - (void)dealloc {
-    NSLog(@"%zd - %s", __LINE__, __func__);
+    NSLog(@"%d - %s", (int)__LINE__, __func__);
 }
 
 @end
